Add accessor function for uio file descriptor
In earlier versions of libuio, interrupt handling routines had to be written
by the application. They have used the fd field in the uio structure to select
or poll on one or more file descriptors, associated with one or more
interrupts.

Later versions of libuio stopped exposing the fields of the uio structure
directly, and introduced accessor functions for the fields that
applications could use. Additionally, a helper function uio_irqwait() was
introduced to allow handling one interrupt without application complexity.

To handle many low-occurence interrupts it is desirable to handle them all
in one poll/select loop. Add an accessor function to obtain the fd field, to
satisfy this use case.

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>