fix missing changes to DEBIAN_FILES
When make distcheck is called the following error is thrown:
make[1]: *** No rule to make target 'debian/libuio-dev.dirs', needed by 'distdir'.  Stop.

The debian package name is libuio2, but this is not chnged in DEBIAN_FILES
variable.

Correct the DEBIAN_FILES to reflect the changes introduced with commit
fa16395873df.

Signed-off-by: Philipp Rosenberger <p.rosenberger@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>