add uio_find_by_base_addr
a helper function that searches the memory map members of each UIO instance
based on a given base address and returns the first UIO info structure that
has the matching base address.

Should be helpful for FPGA based and other embedded systems. Most custom
peripherals are accessed through a base address already but the device tree
that gets generated by tools tend to not use unique names for the field that
UIO uses making it hard to know if the correct peripheral is being used.

Signed-off-by: asm2750 <asm2750@gmail.com>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>