rename to libuio-ng
There are several identically named libuio libraries out there, that
are, as of now, mostly unmaintained. To provide some distinction and
avoid naming conflicts (there is a libuio package in OpenEmbedded/Yocto
already, for example), we add the -ng suffix.

For the sake of compatibiliy, we will not rename any filenames, as it is
unlikely that someone needs to install two different libuio packages on
the same machine (which might name their headers/libraries identically).

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>